chore: refactor event seeder
