+++ test/unit/controller/event-controller.ts
@@ -26,7 +26,6 @@ import EventShift from '../../../src/entity/event/event-shift';
 import EventShiftAnswer, { Availability } from '../../../src/entity/event/event-shift-answer';
 import AssignedRole from '../../../src/entity/rbac/assigned-role';
 import Database from '../../../src/database/database';
-import { seedEvents } from '../../seed-legacy';
 import TokenHandler from '../../../src/authentication/token-handler';
 import Swagger from '../../../src/start/swagger';
 import { json } from 'body-parser';
@@ -43,7 +42,7 @@ import EventService from '../../../src/service/event-service';
 import { EventRequest } from '../../../src/controller/request/event-request';
 import { truncateAllTables } from '../../setup';
 import { finishTestDB } from '../../helpers/test-helpers';
-import { RbacSeeder, UserSeeder } from '../../seed';
+import { EventSeeder, RbacSeeder, UserSeeder } from '../../seed';
 
 describe('EventController', () => {
   let ctx: {
@@ -87,7 +86,7 @@ describe('EventController', () => {
     await User.save(localUser);
 
     const users = await new UserSeeder().seedUsers();
-    const { roleAssignments, events, eventShifts, eventShiftAnswers } = await seedEvents(users);
+    const { roleAssignments, events, eventShifts, eventShiftAnswers } = await new EventSeeder().seedEvents(users);
 
     // start app
     const app = express();
